Ref: Собирает повторяющиеся propTypes в отдельный файл Переносит утилиты и спрайт с иконками
