+++ src/components/main-page/main-page.jsx
@@ -1,17 +1,15 @@
 import React from "react";
 import PropTypes from "prop-types";
-import {Styles} from "../../const";
+import {OfferPropTypes} from "../app/app-prop-types";
+import IconsSprite from "../icons-sprite/icons-sprite";
 import PlacesList from "../places-list/places-list";
-
 
 const MainPage = (props) => {
   const {placesCount, offers, onCardClick} = props;
 
   return (
     <React.Fragment>
-      <div style={Styles.DISPLAY_NONE}>
-        <svg xmlns="http://www.w3.org/2000/svg"><symbol id="icon-arrow-select" viewBox="0 0 7 4"><path fillRule="evenodd" clipRule="evenodd" d="M0 0l3.5 2.813L7 0v1.084L3.5 4 0 1.084V0z"></path></symbol><symbol id="icon-bookmark" viewBox="0 0 17 18"><path d="M3.993 2.185l.017-.092V2c0-.554.449-1 .99-1h10c.522 0 .957.41.997.923l-2.736 14.59-4.814-2.407-.39-.195-.408.153L1.31 16.44 3.993 2.185z"></path></symbol><symbol id="icon-star" viewBox="0 0 13 12"><path fillRule="evenodd" clipRule="evenodd" d="M6.5 9.644L10.517 12 9.451 7.56 13 4.573l-4.674-.386L6.5 0 4.673 4.187 0 4.573 3.549 7.56 2.483 12 6.5 9.644z"></path></symbol></svg>
-      </div>
+      <IconsSprite />
 
       <div className="page page--gray page--main">
 
@@ -123,7 +121,7 @@ const MainPage = (props) => {
 
 MainPage.propTypes = {
   placesCount: PropTypes.number.isRequired,
-  offers: PropTypes.array.isRequired,
+  offers: PropTypes.arrayOf(OfferPropTypes).isRequired,
   onCardClick: PropTypes.func.isRequired,
 };
