+++ src/components/place-card-near/place-card-near.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import PropTypes from "prop-types";
-import {convertRatingToStyle} from "../../const";
+import {OfferPropTypes} from "../app/app-prop-types";
+import {convertRatingToStyle} from "../../utils";
 
 const PlaceCardNear = (props) => {
   const {offer} = props;
@@ -49,15 +49,7 @@ const PlaceCardNear = (props) => {
 };
 
 PlaceCardNear.propTypes = {
-  offer: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    images: PropTypes.arrayOf(PropTypes.string.isRequired),
-    name: PropTypes.string.isRequired,
-    isBookmarked: PropTypes.bool.isRequired,
-    rating: PropTypes.number.isRequired,
-    type: PropTypes.string.isRequired,
-    price: PropTypes.number.isRequired,
-  }).isRequired,
+  offer: OfferPropTypes.isRequired,
 };
 
 export default PlaceCardNear;
