+++ src/components/places-list/places-list.jsx
@@ -1,5 +1,7 @@
 import React, {PureComponent} from "react";
 import PropTypes from "prop-types";
+import {OfferPropTypes} from "../app/app-prop-types";
+
 import PlaceCard from "../place-card/place-card";
 
 class PlacesList extends PureComponent {
@@ -36,10 +38,7 @@ class PlacesList extends PureComponent {
 }
 
 PlacesList.propTypes = {
-  offers: PropTypes.arrayOf(PropTypes.shape({
-    id: PropTypes.string.isRequired,
-  })).isRequired,
-
+  offers: PropTypes.arrayOf(OfferPropTypes).isRequired,
   onCardClick: PropTypes.func.isRequired,
 };
