+++ src/components/reviews-list/reviews-list.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
+import {ReviewPropTypes} from "../app/app-prop-types";
 import Review from "../review/review";
 
 const ReviewsList = (props) => {
@@ -21,7 +22,7 @@ const ReviewsList = (props) => {
 };
 
 ReviewsList.propTypes = {
-  reviews: PropTypes.array.isRequired,
+  reviews: PropTypes.arrayOf(ReviewPropTypes).isRequired,
 };
 
 export default ReviewsList;
